net: stmmac: Disable auto negotiation in QED implmentation
If we enable the MAC2MAC, we instruct mac not to connect and init
the external PHY.
Ethernet expects the other side of SGMII link to act as PHY in terms
of autonegotiation. If the other end is actually a peer MAC, the local
end doesn't receive what it expects in autoneg type SGMII. the link will
be failed.

Signed-off-by: Jason Chen <jaschen@lantronix.com>
Change-Id: Iff3b5b5f7bd428476e081004fce5b48646048307
Signed-off-by: Anthony Wong <anthony.wong@canonical.com>